Fix clearing fetch FIFO twice when CJALR fails
When CJALR fails, for example on a permit execute violation, the fetch
FIFO is cleared twice, first because `branch_req_spec_raw` is asserted
in the `cheri_ex` module and second because `cheri_wb_err_o` is asserted
the next cycle. This causes the first instruction of the trap handler to
be cleared and causes the PC and instruction data to get out of sync.

The solution in this commit is to not assert `branch_req_spec_raw` when
there is an instruction fault.