Formal verification flow for CHERIoT Ibex
This commit adds the content of the dv/formal folder. Please look at the
README in there for more details on how to use this. Here's a bullet
list of what is included in this commit:
- spec_api that is capability aware
- Pipeline follower.
- Comparison utilities between Sail and Ibex.
- Data type invariant for CHERIoT capabilities.
- Alternative LSUs so that we don't need an upper bound on memory response
  time.
- Temporal safety, enabling both trvk and tbre (although revocation
  engine is never started).
- Nix development environment.
- Interrupt verification including non-maskable and fast ones.

The majority of properties (not all) prove with Hp and should be run
individually, apart from Step0 and various specific examples. Some prove
at a higher proof effort than I expect.

Some properties are somewhat problematic, but this seems to be a tool bug.
Querying an SST trace of length 5 fails to find anything, but Hp cannot
prove it. Further, prove -sst 5 finds a trace, but attempting to
load it gives a warning saying an assumption has been violated,
meaning the trace was not a trace at all. These properties do
prove but slower.

Co-authored-by: Harry Callahan <hcallahan@lowrisc.org>
Co-authored-by: Marno van der Maas <mvdmaas+git@lowrisc.org>