fcov: refine ignore bins for bound_check_cases
Since there are starting to be a lot of patterns that we want to ignore,
`define a macro for use at the coverpoints that use this function.

While here, trim the fcov_seal_bound_case ignore bins to the right
width, since it is a slice of the bound_check_cases vector.