fcov: decouple cs[12]_sealed coverpoints and tag
The cp_cs[12]_sealed coverpoints are used in cross points with the
associated tag, and so having these coverpoints also depend on the tag
value results in impossible points in the cross product space.  Drop the
tags from the coverpoint definition, but introduce new cross points,
because the interaction is indeed likely of interest.

Fixes: https://github.com/microsoft/cheriot-ibex/issues/68